mcp-adapters/pyproject.toml
[project]
name = "mcp-adapters"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-mcp-adapters>=0.1.9",
    "langchain[openai]>=0.3.26",
    "langgraph>=0.5.3",
    "openai>=1.95.1",
    "python-dotenv>=1.1.1",
]
